Move to intrin.py
